Combine CodeAction tests in one place
